Updated default tab size to 4 characters from 8 characters as it was leading to incorrect column pointers for the majority of code. The Java standard indent is 4 characters.
Need to add a switch into the options to control the tab size from the script itself. Realistically, 
the number should always be one as the tab is just one character always and the use of tab widths
as far as JavaCC is concerned is just for calculating the position in a column.