Merged in Sreenivasa's (mostly cpp) changes from the version_60 branch (as of revision 559) (since the previous merge which took place at the time of tagging 6.0).
This should now be a superset of Chris Ainsley's changes for 6.1 (Java enhancements plus partial elimination of literal string use), and Sreenivasa's changes for 6.1 (cpp fixes).

The changes as of this commit are entirely Sreenivasa's, but I modified some method names / constant names in order to be compatible with my own changes.