Added the following options :
  PARSER_CONSTRUCTOR_VISIBILITY = "private | protected | public";
  PARSER_REINIT_VISIBILITY      = "public | protected | public";

This allows the parser to have private constructors. This is very useful for when trying to control access to the parser via a static method in the same package as the parser.

The permissions are Java-centric, but are easily adaptable into the C++ code generation. No C++ code generation was altered as part of this checking, and backwards compatibility is maintained. If these options are not specified, then the current behaviour (public constructors and ReInit methods) is unaltered.

If PARSER_CONSTRUCTOR_VISIBILITY is provided without PARSER_REINIT_VISIBILITY, then the PARSER_REINIT_VISIBILITY permission will also apply to the ReInit methods. If both are supplied, then permissions are granular. If PARSER_REINIT_VISIBILITY is supplied by itself, then constructor visibility is unaltered.

Essentially :

ReInit EXTENDS Constructor visibility.