Add missing fork=true to Ant-based tests
Note: almost all Parser-like tests had fork=true, which makes sense.
The Parser class uses System.in to consume the input, and it
is not a good idea to let non-forked **test** code to manipulate with System.in.

The workaround so far is to add the missing fork=true, so parser executions
do not result in System.setIn(..)

See https://github.com/gradle/gradle/issues/12008