Incorrect handling of UNICODE characters in c++ generated code
Between lathe Java language and the C++ language there is a major difference regarding the char data type. In Java this is a `signed` data type with numerical values `0...255` but in C++ this is an `unsigned` quantity with numerical values `-128...127`.
In the code this means that the original line 2951 would also kick in for the unicode values (typical having the high bit set) instead of the code created originally at 2966 (the `else`).
Also the handling of the `hiByte` / `i1` / `i2` as of original line 1674 has to be corrected so that the shifts are done on unsigned values.

As far as I could tell the corrected places are the only places where the problem occurs.

Disclaimer: I have not been able to test the actual Java code due to a lack of infrastructure, I did test the results by adjusting the generated C++ code in the original project